added checkswitch script and updated shell script to check
if switch is on, will not turn off after closing final call in startup.sh

+++ Pi Scripts/startup.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
 
-sudo python MainScript.py
-sudo shutdown -h
+turnOff=$(sudo python checkSwitch.py)
+
+sudo python main.py
+
+if [turnOff = "1"]
+	sudo shutdown -h
+fi
